Move CustomCursor to bevy_winit/src/custom_cursor.rs
